requirements.txt
Alfred-Workflow==1.36
pendulum==2.0.4
python-dateutil==2.7.5
pytzdata==2018.7
six==1.12.0
typing==3.6.6
